duplicate libcontainers cpu share conversion for testing
allows us to trivially support all platforms for the test binary

this conversion ideally needs to be stable anyhow to avoid test incompatibility ...

https://github.com/opencontainers/cgroups/issues/18
https://github.com/kubernetes/kubernetes/issues/135214
